chore: Checkout page changes for build.
